feat(web, server) : customerController, customerService에 설정해둔 auth 기능을 authController, userService에 이동 #205
